requirements.txt
fastapi
uvicorn
pydantic
PyYAML
requests
# LangChain Core & Ollama
langchain
langchain-ollama
langchain-community
# Cloud Provider Libraries
langchain-openai
langchain-google-genai
langchain-anthropic
# For Agents
langchainhub
